fix: swap toRemove & toAdd in deltaExtensions
When we try to remove & add the same extension in a very short
timeframe, it does not get initialized properly because it gets both
added to `toRemove` and `toAdd`. When that happens, the extension is
first added, and then immediately removed. By swapping `toRemove` and
`toAdd` we ensure that the extension is first removed, and then newly
added.