feat: expose `WorkspaceService`
We sometimes want to reinitialize the workspace (e.g. when the user
decides to change their folders). For this, we need to run:

```ts
const workspaceService = accessor.get(IWorkspaceContextService) as WorkspaceService
```

So we can call `initialize` on it. To be able to do this, we need to be
able to access the `WorkspaceService` type.

In this case I decided to expose it from `configuration.ts`, since it's
only used when this service overridden. But let me know if you'd prefer
to expose this from services!